requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --universal
-e .
    # via -r requirements.in
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via sharp
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
fonttools==4.61.0
    # via matplotlib
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
gsplat==1.5.3
    # via sharp
hf-xet==1.1.10 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
huggingface-hub==0.35.3
    # via timm
idna==3.10
    # via requests
imageio==2.37.0
    # via sharp
imageio-ffmpeg==0.6.0
    # via imageio
jaxtyping==0.3.3
    # via gsplat
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.6
    # via sharp
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
ninja==1.13.0
    # via gsplat
numpy==2.3.3
    # via
    #   contourpy
    #   gsplat
    #   imageio
    #   matplotlib
    #   plyfile
    #   scipy
    #   torchvision
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nccl-cu12==2.27.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
packaging==25.0
    # via
    #   huggingface-hub
    #   matplotlib
pillow==11.3.0
    # via
    #   imageio
    #   matplotlib
    #   pillow-heif
    #   torchvision
pillow-heif==1.1.1
    # via sharp
plyfile==1.1.2
    # via sharp
psutil==7.1.0
    # via imageio
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   timm
requests==2.32.5
    # via huggingface-hub
rich==14.1.0
    # via gsplat
safetensors==0.6.2
    # via timm
scipy==1.16.2
    # via sharp
setuptools==80.9.0
    # via
    #   torch
    #   triton
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via torch
timm==1.0.20
    # via sharp
torch==2.8.0
    # via
    #   gsplat
    #   sharp
    #   timm
    #   torchvision
torchvision==0.23.0
    # via
    #   sharp
    #   timm
tqdm==4.67.1
    # via huggingface-hub
triton==3.4.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.6.0
    # via requests
wadler-lindig==0.1.7
    # via jaxtyping



py360convert
Pillow 
piexif